@@ -2,8 +2,7 @@
 using SocialGuard.YC.Services.Security;
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.CommandsNext;
-
-
+using System;
 
 namespace SocialGuard.YC.Modules
 {
@@ -11,30 +10,31 @@ namespace SocialGuard.YC.Modules
 	{
 		public class EncryptionModule : BaseCommandModule
 		{
-			private readonly EncryptionService cipher;
+			private readonly IEncryptionService _cipher;
 
-			public EncryptionModule(EncryptionService cipher)
+			public EncryptionModule(IEncryptionService cipher)
 			{
-				this.cipher = cipher;
+				this._cipher = cipher;
 			}
 
 
 			[Command("encrypt"), RequireOwner]
 			public async Task EncryptAsync(CommandContext context, [RemainingText] string value)
 			{
-				await context.RespondAsync(cipher.Encrypt(value));
+				await context.RespondAsync(_cipher.Encrypt(value));
 			}
 
 			[Command("decrypt"), RequireOwner]
 			public async Task DecryptAsync(CommandContext context, [RemainingText] string value)
 			{
-				await context.RespondAsync(cipher.Decrypt(value));
+				await context.RespondAsync(_cipher.Decrypt(value));
 			}
 #if DEBUG
 			[Command("key"), RequireOwner]
-			public async Task DecryptAsync(CommandContext context)
+			public Task DecryptAsync(CommandContext context)
 			{
-				await context.RespondAsync(cipher.GetKey());
+				throw new NotSupportedException();
+				//await context.RespondAsync(cipher.GetKey());
 			}
 #endif
 		}
